More work on Final
Got the map system sorted out (for the most part). Still temp graphics,
but it's laid out much better, and it'll be MUCH easier to change it in
the future. It's now pulled from an array, so if you use the Tiled
program I talked about, you can export it and get the array from it.
Working on changing the car controls, pulling some code from the
examples in the book to start with. I got it moving on the screen, but
it doesn't change the screen position, which means I may end up removing
Box2d altogether and just rewriting the screen scrolling code. If you
get a chance, try integrating a target reticule and some kind of basic
shooting mechanism from the car. Should be easy enough to pull some old
code for that.