Audio and Image additions, plus guns?
Got the sprite sheet working for background stuff. I'm having an issue
with flickering though, so I'm not sure what to do on that end. I'll
have to explore it some more. Added basic audio support, still tinkering
with that. Added a cursor and currently working on a basic gun setup. I
ripped straight from my midterm, but it's not yet working, still
tinkering with that.